move graphql-mocks
