+++ engine/crates/graphql-mocks/README.md
@@ -1,0 +1,3 @@
+A collection of GraphQL Mock schemas, along with an axum server to run them.
+
+Useful for integration tests of federation & the GraphQL connector.
