+++ engine/crates/integration-tests/tests/federation/basic/fragments.rs
@@ -1,5 +1,6 @@
 use gateway_v2::Gateway;
-use integration_tests::{federation::GatewayV2Ext, mocks::graphql::FakeGithubSchema, runtime, MockGraphQlServer};
+use graphql_mocks::{FakeGithubSchema, MockGraphQlServer};
+use integration_tests::{federation::GatewayV2Ext, runtime};
 
 #[test]
 fn named_fragment_on_object() {
