+++ engine/crates/integration-tests/tests/federation/basic/headers.rs
@@ -1,7 +1,8 @@
 //! Tests of header forwarding behaviour
 
 use gateway_v2::Gateway;
-use integration_tests::{federation::GatewayV2Ext, mocks::graphql::FakeGithubSchema, runtime, MockGraphQlServer};
+use graphql_mocks::{FakeGithubSchema, MockGraphQlServer};
+use integration_tests::{federation::GatewayV2Ext, runtime};
 
 #[test]
 fn test_default_headers() {
