+++ engine/crates/integration-tests/tests/federation/basic/streaming.rs
@@ -4,7 +4,8 @@
 
 use futures::StreamExt;
 use gateway_v2::Gateway;
-use integration_tests::{federation::GatewayV2Ext, mocks::graphql::StateMutationSchema, runtime, MockGraphQlServer};
+use graphql_mocks::{MockGraphQlServer, StateMutationSchema};
+use integration_tests::{federation::GatewayV2Ext, runtime};
 
 #[test]
 fn can_run_a_query_via_execute_stream() {
