+++ engine/crates/integration-tests/tests/graphql_connector/defer.rs
@@ -1,4 +1,5 @@
-use integration_tests::{mocks::graphql::FakeGithubSchema, runtime, EngineBuilder, MockGraphQlServer, ResponseExt};
+use graphql_mocks::{FakeGithubSchema, MockGraphQlServer};
+use integration_tests::{runtime, EngineBuilder, ResponseExt};
 
 #[test]
 fn test_defer_on_graphql_connector() {
